python/requirements.txt
pandas~=1.2.1
kafka-python
pymongo~=3.11.2
dnspython